refactor(core): StringEncoding
