+++ .licenserc.yaml
@@ -72,7 +72,6 @@ header: # `header` section is configurations for source codes license header.
     - 'assembly/**'
     - '.github/**/*'
     - '**/target/*'
-    - '**/util/StringEncoding.java'
     - '**/go.mod'
     - '**/go.sum'
   comment: on-failure # on what condition license-eye will comment on the pull request, `on-failure`, `always`, `never`.
